requirements.txt
fastapi==0.75.0
SQLAlchemy==1.4.32
uvicorn==0.17.6
starlette==0.17.1
python-jose
passlib[bcrypt]~=1.7.4
pydantic~=1.9.0